% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_abc}
\alias{rc_abc}
\title{Create a vector of string using C}
\usage{
rc_abc()
}
\description{
Keywords: STRSXP, allocVector, PROTECT, SEXP, SET_STRING_ELT, mkChar, length,
STRING_ELT, CHAR, const char *, UNPROTECT, puts, type2char, TYPEOF, asChar
}
\details{
puts function is to print a char array (or string).
}
\examples{
\dontrun{
rc_abc()
}

}
