% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_add}
\alias{rc_add}
\title{Return a sum of two numbers}
\usage{
rc_add(a, b)
}
\arguments{
\item{x}{description}
}
\description{
Keywords in the C code: REALSXP, allocVector, PROTECT, SEXP, asReal,
UNPROTECT
}
\examples{
\dontrun{
rc_add(1, 5)
}

}
