% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_add_four}
\alias{rc_add_four}
\title{add 4 to a number in place with duplicate function in C}
\usage{
rc_add_four(x)
}
\arguments{
\item{x}{the value to be changed in place}
}
\description{
Keywords: duplicate, PROTECTSEXP, REAL, UNPROTECT
}
\examples{
\dontrun{
x <- 1
y <- x
rc_add_four(x)
y
x
}
}
