% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_add_two}
\alias{rc_add_two}
\title{add 2 to a vector}
\usage{
rc_add_two(x)
}
\description{
Keywords: length, double *, REALSXP, allocVector, PROTECT, SEXP, REAL,
UNPROTECT, for
}
\examples{
\dontrun{
(x = rnorm(5))
rc_add_two(x)
}

}
