% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_coaerce}
\alias{rc_coaerce}
\title{coaerce a vector to numeric}
\usage{
rc_coaerce(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a list of two, which are the coaerced vector and a logical to tell
if the original vector is numeric (double).
}
\description{
Keywords: ScalarLogical, PROTECT, TYPEOF, coerceVector, REALSXP, SEXP, UNPROTECT,
SET_STRING_ELT, const char *nm[], VECSXP, length,
allocVector, mkChar, install, setAttrib, SET_VECTOR_ELT
}
\examples{
\dontrun{
rc_coaerce(as.numeric(1:4))
rc_coaerce(1:4)
rc_coaerce(c(TRUE, FALSE))
}

}
