% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_createList}
\alias{rc_createList}
\title{create a random list of 3 elements}
\usage{
rc_createList(x)
}
\description{
Keywords in the C code: REALSXP, LGLSXP, INTSXP, asInteger, PROTECT,
UNPROTECT, allocVector, SET_VECTOR_ELT, SEXP
}
\examples{
\dontrun{
rc_createList(4)
}

}
