% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_get_dim}
\alias{rc_get_dim}
\title{get the dimension of a matrix}
\usage{
rc_get_dim(obj)
}
\arguments{
\item{obj}{a matrix}
}
\description{
Keywords: GET_DIM, duplicate, PROTECT, SEXP, UNPROTECT
}
\examples{
\dontrun{
rc_get_dim(matrix(1:10, nrow = 5))
rc_get_dim2(matrix(1:10, nrow = 5), attr = "dim") # here it is not R_DimSymbol
}

}
