% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_is_na}
\alias{rc_is_na}
\title{Are the elements in a vector a NA}
\usage{
rc_is_na(x)
}
\description{
Keywords in the C code: for, switch, TYPEOF, NA_LOGICAL, LOGICAL,
INTEGER, NA_INTEGER, ISNA, REAL, NA_STRING
}
\examples{
\dontrun{
rc_is_na(c(1, 2, 3, NA, 4))
rc_is_na(c(NA, "a"))
rc_is_na(c(NA, TRUE))
}

}
