% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_length}
\alias{rc_length}
\title{length of a vector}
\usage{
rc_length(x)
}
\description{
Keywords: XLENGTH, R_xlen_t, ScalarReal, (double), PROTECT, SEXP, UNPROTECT
if you use `int n = length(x)`, then this function cannot deal with a vector
of length 2^31 -1.
}
\examples{
\dontrun{
rc_length(1:5)
rc_length(1:2^32)
}

}
