% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_make_matrix}
\alias{rc_make_matrix}
\title{make a matrix}
\usage{
rc_make_matrix(obj, nrow, ncol)
}
\arguments{
\item{obj}{vector of integer or numeric}

\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\description{
Keywords: asInteger, allocMatrix, REALSXP, TYPEOF, if, else if, else, PROTECT, double*,
REAL, for, UNPROTECT, INTEGER, int*, UNPROTECT, error
}
\examples{
\dontrun{
x <- 1:10
(y = rc_make_matrix((x), nrow = 5, ncol = 2))
class(y[1])
(y = rc_make_matrix(as.numeric(x), nrow = 5, ncol = 2))
class(y[1])
}

}
