% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_new_call}
\alias{rc_new_call}
\title{make a call object `10+5`}
\usage{
rc_new_call()
}
\description{
Keywords: ScalarReal, PROTECT, SEXP, install, LCONS, R_NilValue, UNPROTECT
Here install function is to create a symbol (the equivalent of as.symbol() in R).
}
\examples{
\dontrun{
rc_new_call()
}

}
