% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_set_attr}
\alias{rc_set_attr}
\title{set attribute for an object}
\usage{
rc_set_attr(obj, attr, value)
}
\description{
Keywords: asChar, CHAR, duplicate, setAttrib, install, SEXP
}
\examples{
\dontrun{
x <- 1:10
rc_set_attr(x, "a", 1)
}

}
