% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_tag}
\alias{rc_tag}
\title{get data from pairlist}
\usage{
rc_tag(x)
}
\description{
Keywords: TAG, CAR, CAAR, CDR, CADR, SEXP, UNPROTECT
if you use `int n = length(x)`, then this function cannot deal with a vector
of length 2^31 -1.
}
\examples{
\dontrun{
x <- quote(f(a = 1, b = 2, c = 3, d = 4, e = 5, f = 6))
rc_tag(x)
rc_car(x)   # first
rc_caar(x)
rc_cdr(x)
rc_cadr(x)  # second
rc_cddr(x)
rc_cdddr(x)
rc_caddr(x) # third
rc_cadddr(x)# fourth
rc_cad4r(x) # fifth
rc_car(rc_cdr(rc_cdr(rc_cdr(rc_cdr(rc_cdr(x)))))) # sixth
rc_car(rc_cdr(rc_cdr(rc_cdr(rc_cdr(rc_cdr(rc_cdr(x))))))) # seventh
}

}
