% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCwithInline.R
\name{rc_zeroes}
\alias{rc_zeroes}
\title{create a vector of zeros}
\usage{
rc_zeroes(n_)
}
\description{
loop through each element in the vector and set it to a constant. The most efficient way to do that is to use memset()
// void \* memset ( void \* ptr, int value, size_t num );
Sets the first num bytes of the block of memory pointed by ptr to the specified value .
please not the character for new line is no longer `\n` but `\\n`.
}
\details{
Keywords in the C code: INTSXP, asInteger, allocVector, PROTECT, memset,
UNPROTECT, for, printf, INTEGER
}
\examples{
\dontrun{
rc_zeroes(10)
}

}
