% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCAR_CDR.R
\name{testCAR_CDR_ext}
\alias{testCAR_CDR_ext}
\title{test CAR and CDR macros in C}
\usage{
testCAR_CDR_ext(a, b)
}
\arguments{
\item{a}{first parameter}

\item{b}{second parameter}

\item{c}{third parameter}

\item{d}{fourth parameter}

\item{e}{fifth parameter}

\item{...}{the rest parameters}
}
\description{
test CAR and CDR macros in C
}
\examples{
\dontrun{
testCAR_CDR(1,2,3,4,5)

testCAR_CDR(1,2,3,4,5,6,7,8)
}

}
