% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFile.R
\name{writeFile}
\alias{writeFile}
\title{writeFile}
\usage{
writeFile(fileName, num)
}
\arguments{
\item{fileName}{fileName}

\item{num}{the numbers to be saved in the file}
}
\description{
writeFile
}
\examples{
\dontrun{
writeFile("./tmp.txt", 1:10)
}
}
